Fix issues related to disabling the server
First issue was a crash when disabling the server and switching the title. If the server was enabled, then the server was disabled in the config menu, changing the title to something different would cause a crash in the plugin.

Second issue is the reason why the first issue happened: when the server was disabled in the config menu, the server did not actually get disabled. This fixes that.